sm6150-common: overlay: Enable UDFPS framework dimming
This is currently only supported on a handful of ROMs, but this allows the framework to do the screen dimming for optical in-screen fingerprint sensors instead of the kernel (which our device doesn't support).

The calculations are based from the Lineage 17.1 FOD HALs used for Samsung Galaxy As[1] and may be off, but these can be tuned once we tested this.

[1]: https://github.com/erfanoabdi/android_device_samsung_a51/blob/lineage-17.1/fod/FingerprintInscreen.cpp

Change-Id: Iddeecbb45e3d62a3df4042ad5bca6beb6f5acfb1